fix: only spawn node once
